A first cut of a webpack config.
The goal of using webpack to bundle the extension is to make
it easy to switch to typescript and break up the code into
multiple files for easier development for everyone.

This initial configuration, inspired by the example on github
in the chrome-extension-webpack repo, does very little other
than linting with eslint and bundling with babel.

It'll be the base from which we switch to the actual linter
I want to try out (rome) and convert the code to typescript
as desired.

A variety of eslint errors had to be fixed in intel.js to
make this change work. However with that done, it is possible
to run

npm run start

to watch all the source code and automatically build the dist
directory, and then to load dist as an unpacked extension to
test the development version.