Remove noDoubleEquals rule from linter for now.
Neptune's Pride uses a lot of strings for numbers, and === is
error prone because it makes tests for things like fleet or userids
fail as the types mismatch in some cases.

For now, just disable this rule and revert all of its fixups, even
those that were OK. Fixes bugs in combat calc and in star report.