Extract image_url into its own file by way of example.
Here is a minimal example of what you need to do to move
parts of intel.js into a new file.

It's as straightforward as cutting and pasting the original
into the new file and putting the keyword 'export' before each
declaration you want access to outside that file. Then, in
intel.js, you can import the symbol and use it as before.