Make a slightly more logical and isolated hotkey library.
Though this isn't really much improved, change the naming and
interfaces for hotkey definition and isolate the code into a single
file where it is central so that it can be used by all features of
SAT instead of only the original NPA features. The hope being that
this new interface makes enough sense and is re-usable so that other
feature authors will want to use it to hook up their hotkeys.

Documentation is now passed into the new defineHotkey method, so
that when you want to hook up a function to a hotkey you can merely
call:

defineHotkey(myKey, myFunction, myHelp);

and the resulting hotkey will be wired up and help will appear when
the user types '?' in the UI.

I ran into a problem when writing this change, which is that the
js validation is really way too loose. It is easy to move some
variable out of intel.js and not import it, resulting in undefineds
that are not caught by the linter because it expects them to be
caught by the typescript compiler.

For this reason I will next extract all of NeptunesPrideAgent code
into its own typescript file, which will make all further refactoring
of that code less error prone.