Add a README.md and move source files to prep for v2.
For version 2 of the extension, I want to write it in typescript
with real javascript tooling for a vastly superior and potentially
much more collaborative development setup.

I have permission from Jay of Iron Helmet Games to publish this
extension as open source and use his assets to advertise it on
the store as long as it's clear to players that this extension
and software derived from it are not the game itself. So as of
this commit, the repository will be in the open where others
can also collaborate.