Implement [[Tick #n]] substitution for better times.
The new [[Tick #n]] substitution can give you times in ETA in
your local timezone, relative real time, relative tick time,
and absolute real time.

It defaults to ETA and affects all times in the UI, and
the entire UI now refreshes when the user presses the % key
to cycle between the options.