Fix long-standing autocomplete issue for fast typists.
If you typed [[0] fast enough, the autocomplete event handler
would be called with ' [' and then '[0', causing the start
position not to be detected and autocomplete would not trigger.

Instead, check for completion on every ']', ignoring cases where
the completion was already done or there isn't even an '[[' in
the text at all. Now autocomplete always completes if it's possible
to do so.