Add ability to recolor players, and w hotkey.
Generically make the code respect the player color given
by the values in NeptunesPride.universe.galaxy.players[uid].color
and add a recolorPlayers() function that will reset all sprites,
CSS, and images to the colors as specified.

This has the side effect of changing the size of any 48x48 ownership
rings, but these are still small enough to appear OK in the UI.

By default, other than that size change no change is apparent. If
the user uses the 'w' hotkey, their own colour will be changed to
white so that they can easily distinguish their own empire on the
map.

As a follow up, it would be nice to be able to persistently set
any player's colour.