Check if we are on Chrome before overriding touch.
Touch handling and mouse handling works fine on Chrome without
Jay's original code to distinguish cases, and for touchscreen
Chromebooks disabling his original code enables the trackpad
and touchscreen to work in tandem.

On iOS Safari, this would disable touch handling in favour of
a trackpad that the user probably does not have, so guard this
as a Chrome-only hack so that the iOS version of the extension
works.

Since I don't know what happens on firefox, leave the default
behaviour in place.