Make time display persistent and have ticks update UI.
Make the default time display relative (i.e. the ticking production
clock as usual in the game) but also make the '%' key store the
user's preferred time setting persistently so that if they change
how the clock works it sticks across sessions and games.

Also watch the clock for the 29th second, so that times formatted
from [[Tick #]] will update to the correct minute every time the
rounded time would adjust itself; this makes the map repaint and
the UI refresh. This seems mostly good but it can be annoying if
you are in a long message thread because it will focus the comment
box. Let's see how much people hate that.