Player recolouring was completely wrong for 64p games.
First off, the CSS for the player underbar doesn't exist and was
throwing an exception and completely breaking the code for 64p
games.

Secondly, there were not enough glow icons in the sprite for
every player to have a unique glow. Make the destination sprite
bigger to handle the extra positions, draw the extra glows, and
override the sprite positioning for stars with gates. The end
effect is unique glows per player that correspond to their sprite.