Improve the behaviour of the NP Agent menu item.
First off, multiplex the item with the new fleet item
so that it gets the benefit of being a "legit" screen
according to the rest of the UI - namely, that it will
be refreshed automatically if the UI is refreshed (for
example if the user changes timebase with '%').

Second, make it remember what report you were on so that
it can show you the report you wanted. Very likely users
will often want to see the ledger or API key report over
and over again.

Third, position it above the back menu item.

The code is ugly as sin but the UX is great.