Make a hotkey to bring up the UI.
The UI itself is useful enough that the user might actually want
to use it, especially for the API Key screen or the short fleet
reports or the accounting.

Bind ` to bring up the UI and make it so that if you clip a report
it is the default view (otherwise it is the last thing you looked
at in the UI). So for example k` brings up API keys, a` brings up
the accounting, ^` brings up the short fleet report.