Add the ability to link players with only their symbol.
You can already type [[0]] to link Player 0 and display
their name as a link alongside their symbol; add [[#0]]
to do the same but with only their symbol as the link.

In non-extension games this will show up as (#0) which
is hopefully clear enough, and in extension games you
can now make dense tables with the player's symbol as
the column heading.