Add the ability to make NPA buttons.
Now the text [[hotkey:<key>]] or [[goto:<key>]] will create a
button that can trigger an NPA action. Normally you will want
[[goto:<key>]] so that the user sees the UI corresponding to
the report; the hotkey one is just like they pressed the hotkey
which in many cases just puts stuff on the clipboard and doesn't
navigate.

Use this to create a crude controls table that displays all the
hotkeys so that users who are on phones or tablets without keyboards
can find and execute all the functionality in NPA.