Fix merge all keys when a bad key is in the list.
If an ally has regenerated their key so it no longer works, make
their recorded key "badkey" so that the user can tell there is a
problem and go fix it.

Also fixes the problem that merging all would fail for all in this
case, and fixes a race condition in setting the keys in the first
place.