Fixes for turn based games.
Put the submit button in its proper place so it isn't overlapping
all the screens.

Change "eta" time format to not use clock times, which are meaningless
since the clock is entirely driven by the players completing their
turns. Instead, make ETA in terms of integer turns so that you can
know how many turns away everything is and use that for planning.