Make table rendering pickier about the header line.
Some players actually use --- in their messages, which causes
the table render to trigger. Make it dramatically pickier so
that table rendering is only invoked when definitely intended.

Now --- must be at the very start of the line, and a second ---
must be in the line, in order to decide you want a table following.