Complete research screen for allies.
Provide a research report (hotkey E) that summarizes research
for everyone in the alliance based on API keys and contains
the live data from the other players.

The resulting report can be emailed and times will update, but
the times will be based on the historical research rate at time
of paste. Still useful for getting live updated ETAs, but one
must be careful about the distinction between getting a fresh
report with everyone's current science and seeing a potentially
stale report based on science as at a certain time.