Two fixes: pickiness update, and immediate format.
My previous attempt to be pickier about --- occurrances was not
correct, and is now hopefully fixed.

The <br> tags inserted into comments are different on the client
side than on the server side, update the split to use a regex that
works on both so that things are properly formatted the first time
round (notably tables).