Fix wrong event wiring so screens aren't re-executed.
The event handling for the reports screen was accumulating at the
top level of the UI causing the screens to be rerun as many times
as the UI has previously been displayed. For most screens this was
fast enough to be close to imperceptible but for the new heavy
reporting screens they get decidedly sluggish after being in the
client for a while.

Hooking up the event properly means calling the report exactly once
and is pretty snappy.