Fix fleet merge bug with multiple allies.
If multiple allies could see a fleet then merging would take the
last allies' view of the fleet or your own view of the fleet in
preference to the scan data of the owning player. Fix it so that
if you have a scan from the owner, that view takes precedence over
all others, providing the detailed scan data with a full order set.