API Key convenience features.
When the user generates an API key automatically cache their own
scan data so that the API key will forever be known to this instance
of NPA.

On every UI refresh (which we are doing every minute) ensue that the
allies API data are fresh in the cache. Never wait again for the
research screen to load.

Inside message compose, autocomplete [[api: with the user's own API
key. If they really wanted to send someone else's, they can go back and
paste it over their own key.