First iteration of the auto ruler.
The auto ruler shows you the nearest enemy attacker of a selected
star and the nearest defender by default. If the defending star
can see the attack in time to launch defense its defense line is
shown in green, but if it can't make it in time its defense line
is shown in grey.

Pressing 8 will turn auto ruler power down to 0 and show nothing.
Pressing 9 will turn up auto ruler power and show you all closer
defending star candidates; and every other level adds another level
of attacker distance.

The setting of the auto ruler is not yet persisted and defaults
to power level 1.