Handle ruler cases where the world is rotated upside down.
In the drawing code when we're flipped the arc for visibility needs
to be drawn on the other side.

Also use real distances for deciding whether to say it's invisible,
rather than drawing distances which are cut short to look nicer and
make longer ranges (9 ticks) look invisible accidentally in the
previous iteration.