Fix a bug where gates weren't being used in the initial sort.
When I switched to a sorting strategy for distances, I failed to
realize that the sortedStarSprites don't contain gate information
so the nearest defender was being found without considering warp
effects. Fixed it to use stars so that the calculations
are as accurate as possible given the current game state.