First working typescript settings creator.
After much futzing, figured out how to convince typescript what
the types of individual settings are and be able to add them on
the fly. They get loaded from the database on initialization and
updated each time they are written, but otherwise function as
locally cached variables throughout for efficiency.

There may be a race condition here and there's a console.error in
case that ever happens, but even if hit the effect is minor for the
one or two settings that exist at the moment. This race might require
more though to get right.