Redo color choices for text.
After some use it just seemed that the red text was never easy
to read, but the grey color for ineffective support was really
helpful. So change it so that all autoruler text is green except
for the ineffective support which continue to use grey. Stays in
the theme of the HUD colouring (same as the scan grey/green
distinction) and makes it easy to tell what support works and
what is ineffective.