Rework how the message cache gets loaded.
Use the new fetch API rather than jQuery for the POST request,
and rearrange the async flow to work properly and prioritize
getting events from the cache.

In the past there was a problem behaviour where the first load
of the accounting screen could be slow or even fail outright. In
the fixed async flow, this should never happen; the accounting
screen should be instantaneous or nearly so in all cases, except
if the player joins a very old game from a new client. Even then,
the player might never experience a slowdown as the events will
get cached when they first join the gaem.

There is some noisy log output for now in case there is a problem
with this change but other than that no negative side effects.