First cut of using npaserver for game history.
Game snapshots are copied into the local database from npaserver
once in bulk at the start, and as they occur thereafter, so that
every client has a complete log of the game changes.

The first report built on top of this is "Star Ownership" which
shows all star ownership transitions and makes it easier to figure
out who is at war with who and who is abandoning stars for who.