Add explorers and activity report.
When star ownership has nothing to show you, show exploration
of neutral stars instead for start of game intel.

Add an activity report that shows the activity timestamps for
the player who is currently selected on the map, or for yourself
if nobody else is selected. Show the tick, alias, and upgrade
status for each active timestamp.