At long last, the time display for time machine looks good.
After endless futzing trying to figure out the right approach,
I think this covers it. Time travel to a prior tick puts you
at the start of that tick, all times in the HUD display correctly
as if you really were at the beginning of the tick from the past
and you can replay and see everything update as you'd expect.

Meanwhile, enhance the time machine timestamp to be special and
know that if you're time traveling you're not really at relative
tick 0 but actually offset from the true tick. So when it is
displaying relative measures it gets to break the illusion and
show you -1h or -1 tick instead of 0 all the time.

The only thing that maybe isn't the display you'd expect is the
time to production. Since you time travel to the _beginning_ of
the historical tick, it's always 59m to production at first,
even if you know the real production time is coming up soon. This
seems OK to get all the other times looking perfectly consistent.