Avoid needless readback of intermediate draws.
Though the eventual starSrc has to be read back in order to avoid
disturbing the rest of the client code, the individual pieces of
the sprite can be much more efficiently drawn directly into the
canvas, so avoid needlessly reading them into an image and painting
them the slow way.

The speedup on the whiteout functionality is marked.