Make time travel more robust in the face of missing ticks.
The server side is having a problem where we are dropping lots of
ticks and the client wasn't prepared for that. Now it more robustly
skips over tick#s where no data are available making it possible to
at least review what you have even if the data store isn't perfect.