Update the star report to handle darkness.
Accumulate keys in the star report from each tick rather than
using only the keys from the first tick, so that as darkness
lifts the keyset gets larger. Also should have the side effect
that if you lose sight of systems, ownership is presumed
unchanged, which will be mostly right (and is the best you can
do as you can't see the activity anyway). If when you regain
visibility ownership has changed, it'll be reported as happening
then.