Use a new strategy for keeping track of colors.
Instead of having different approaches for Proteus and Triton,
have a single palette that maps from the player's UID to their
intended colour and use that everywhere. Use the full_universe
order to set the color mapping correctly if it is a proteus game,
and use the default colour map (hardcoded) for triton games.

At some point, it will be trivial to change it to load this color
map from local storage, so the user can set their own colors for
any player and the approach will still work nicely for that. So
this sets up well for the longer term feature, too.

Of course this wasn't just for fun: before this proteus would recolor
everyone who shared your colour instead of just you, and that was no
good. Now, proteus recolor works properly, too.