Update event/message caching to handle messages.
In the event log the entries are immutable, so it is OK to rely
on the key alone to disambiguate where the overlap is and use set
on the db to verify that each new incoming record is in fact new.

On message threads, the top level event that contains the thread
is mutated each time a new comment comes in. There is an activity
timestamp and a comment count either one of which could be used
to detect this situation. Use the comment count to accept duplicate
diplomacy events and write the with put() so that the database will
update calmly whether it is new or an update.

As a result of this change, the client now has the entire game
event history and the header message of every thread. What remains
is to also cache and index the comment messages to enable a whole
host of messaging based features.