First version of diplomacy caching.
With a probable bug where if you start using NPA mid-game the cache
will exclude comment thread history for threads > 10 messages, this
first implementation will cache all diplomacy in the database so
that it can later be searched for various features.

Built on top of the same infrastructure as the event cache, with
one 'group' per message key, this should work neatly to enable finding
old API key or searching messages to find that hard to find thread.