Build an inverted index on all diplomacy messages.
Build an ivf on diplomacy messages, and as a proof of concept use
it to console.log all the API keys this user has seen to the
javascript console. This can become the basis for recovering from
regenerating your API key for history as well as enabling the
tracking of all shared API key history for the extension, fixing
the one major bug with the time machine at the moment, which is
that it can't handle regenerated API keys, and adding the one major
missing feature, namely the ability to use the time machine with
merged keys.