Make time travel to merge all known scan data for historical ticks.
So control-left and right didn't seem to work so now it's c-,
and c-. as those keys usually have < and > printed above the
punctuation.

Change time travel to find all scans for the desired destination
tick and merge those into the display. This default seems the most
desirable as it shows you all the historical data you have for any
given tick.