Make the time machine move in turn increments.
In turn based games, it is annoying to have to step through multiple
ticks of "missing data" because of the 8 tick turn increment. Instead
read the game configuration and jump around in turn increments so
that you can play back the game as you might expect given that there
are no intermediate updates.