Update pimple drawing to respect range adjustement.
Previously the pimples for borderline stars were based on the
star owner's true range, without the handicap adjustment. Fix
up the drawing to be simpler for star territory so that the
range adjustement would be factored out, and then use the
refactored range calculation for the pimples too so that the
feedback is always right.