Have the message cache read state be authoritative.
The server expects that if you request a comment thread, you must
be on the screen that displays the comment thread because that's
the only way that ever happens in the vanilla game. However since
NPA is caching all the diplomacy for you behind the scenes, that
means that the read state is wrong for the UI.

Update it so that the read state in the database cache is
authoritative, and that read state changes will trigger database
updates. This seems to fix the unread count/unread message problems.