Fix up some bad boundary condition bugs.
In bulk trade, if you didn't have the cash to send a level, the
new code to try to send all levels would wind up in an infinite
loop. Just give up in this case.

In time travel, traveling forwards across an API key reset
boundary would fail to start at the right place in the new block
and give up on finding data. Fix it to start at the beginning
when moving forward, and at the end when moving backwards. Linear
is still fast enough but probably I will want to make this a binary
or galloping search if the tick # is far enough away from the current
index...but at least it is now working well in the presence of
multiple API keys in all directions.