Change territoryBrightness behaviour when display is off.
Sometimes users have turned off territory display with ) by accident
and then try to use ctrl-8 / ctrl-9 to change the style to fix it,
but think it's broken. If display is off and you change the style,
turn it on first (and don't change the style); then if you change
the style again it will change as expected.

Also attempt to make all hotkeys override browser hotkeys by telling
Mousetrap that is what we want, so that c-8 / c-9 don't also switch
tabs on Windows and Linux.