Charge for alliances in the accounting screen.
When you request peace, consider that the other owes you the full
price of peace until they accept. If/when they accept, count that
as half the value so they still owe the other half. On game defaults
that means someone owes you $150 as soon as you offer peace and $75
if they accept.