Fix the activity report to use all available scan data.
Instead of relying on the user to have a single key for the whole
game, search all known API key data to build the activity report.

The approach taken is horrifically inefficient, but after caching
the JSON.parse results the time varies between about 60 - 100ms,
fast enough to be not worth fixing yet.