If the galaxy is already loaded, record the true tick.
In the previous implementation there was a race condition where
the time wouldn't get set if the universe was loaded before the
extension hooked the event. Now check for that and initialize in
either case so that the Tick indicator in the HUD is always correct
and reports that rely on it like star ownership are populated.