Implement Gingey's suggestion for faster time travel.
Gingey suggested it would be great to be able to move back and forth
multiple ticks at once, or to be able to jump to a specific tick.

Added hotkeys control-m and control-/ (these are on either side of
the < and > keys which correspond to control-, and control-.  for
moving the time machine one tick) as "fast" movement, which move
you an entire cycle's worth of ticks in one keystroke (typically
24 or 20 ticks at once).

Also linkify everywhere a time appears in the UI so that you can
click that time and it will immediately warp you to that tick. Great
for jumping to a specific spot in the star ownership report or
activity report. Eventually when forwards time travel is implemented,
will also be good for jumping forward to next production or to a
research complete tick to view the galaxy with the production or
research applied. But that's a dream for another day.