A brainwave on message read state - don't cache unreads.
The big issue with the message indexing/caching is that it
triggers the server side to mark threads read when the user
hasn't read them yet. I really wanted to know all the information
for API key surfacing and search purposes all the time, but
the set of hacks and fixes to try to get the unread state
to be correct just kept building up and up.

Finally the brainwave: if the issue is unread threads getting
marked read without user action, simply don't add those to the
cache until the user _reads_ them. It even makes a sort of sense
that NPA shouldn't show you information from threads you haven't
even opened, and of course the expectation is that everyone will
open every thread over the course of the game so all the useful
information will eventually wind up in cache. And it seems to me
that this guarantees correctness and will enable removal of all
the prior hacks.

Time to test it out and see...