The beginnings of efficiency for the time machine.
The strings of game state alone can add up to about 1G of RAM
in the client, and the parsed game states in that case were 700M,
both just too much RAM enough it runs fine on today's crazy
powerful devices.

Try out an object prototyping approach which ultimately turned
out to be a failure, saving not nearly enough RAM and costing
a ton of execution time. A diff/patch system that can only move
from one game state to an adjacent one is going to be the clear
winner because that's how we navigate game state anyway and the
diffs/patches are dramatically smaller than the entire game state,
often just a timestamp.