First version of patch based time machine.
With Rob's help, implemented two versions of patching and
used the object based one as the basis of a time machine
that no longer stores full states except for the endpoints
of each block of API data. The code now needs dramatically
less RAM and performance is OK except at initial load time
when all the patches get computed.