Patch based time machine v2.
In this version, only patches are saved to the database and the
database is validated at load time and rebuilt if necessary. The
rebuild is quite costly on big games like the 64p game, but on
normal size games it is not really a problem.

This sets the stage for receiving diffs from the server side, which
will be in the same format and require much less bandwidth to sync
not to mention the already huge RAM savings.

This was quite tricky to get working so it may be a few days to get
it out to real alpha users, let alone stable.