A slow but definitely working version that clones scans.
The client code in intel.js was relying on the fact that the
scans were all separate, mutable objects with long lifetimes
and this meant that the new diff-based approach to scan data
made the star ownership and activity reports and long-range
time travel all not work properly.

As the simplest fix, clone the return value of getScan in a
new getScanClone entry point and verify that everything is
now "working"; and it is working, it's just really slow.