Don't hide the UI when time traveling.
It is actually often convenient to have the UI up when moving
through time, for example to watch a player's empire page for
tech changes. So leave whatever UI screen the user had up up
when they time travel.

Unfortunately for NPA's own UI this takes you back to the
relatively useless home planets report. That probably needs
fixing in a future patch.