Update battle calculator to display the tipping point.
Now, for every combat, indicate what the losing side requires in
order to change the outcome, regardless of who is losing. Useful
to see what you need when your attack/defense is lacking, and also
useful to see if the enemy has a hope of defending against you
with the help of nearby systems.