Add filtered combat reports.
Make it possible to see the equivalent of the long fleet report
but only those stanzas which contain a substring of interest. Two
new reports added, one that looks for "Combat!" so that you can
review all combat outcomes, and one that looks for [[NAME]] so you
can look for a specific space object (fleet or star) and see all
the fleet travel projected to affect it.

Ultimately this will be used for notifications to take you directly
to a combat summary of what the notification was about, using the #
portion of the URL to query the filtered fleet summary.