Fix a star data merge bug.
Take the owner's scan as the authoritative one for a star so
that in-flight orders like ship transfers are properly accounted
for in the merged view.

Note that even FA scan information doesn't give you this view, so
you're better off looking at the map with the API key merged than
relying on the FA scan info sharing, which doesn't include fleet
orders and ship transfers.