Update the combat calculator for multi-round FA combat.
If multiple players from a formal alliance landed at once,
defeating an enemy system, the code would then have them
fight it out as usual without realizing that they are allies.

Now, the system is credited to the FA's biggest contributor,
and all incoming FA fleets are added to the star's total
defenses, and combat will continue if any offense remains.

Possibly/probably there will be a bug in cases where the FA
wins initially but then loses to an incoming FA. It's a hard
situation to manufacture.