Fix a nasty bug in the star report caused by corrupt diffs.
The new diff code has a flaw where if either the source object
of the diff or the patched result of a diff are modified, that
could corrupt the diff itself, making repeated diffs produce
incorrect answers. This shows up in the ownership report which
is executed twice due to a different bug with hotkey implementation
and the second execution gets the wrong result, so that the
star report is only reliable if chosen from the GUI dropdown
after a reload.

Add tests for the problem in diff2/patch2, fix the underlying
problem, and fix an object aliasing issue in the star report.