Update formal alliances to know the most recent tick.
In preparation for being able to process declarations of war (which
the code still doesn't do as I don't have a ready exmaple of it
yet) the knownAlliances are recorded with the tick # of the most
recent observed allied activity. So now, if someone declared war
_after_ that, the end of the alliance can be known and used to clear
out the alliance to enable correct calculations.

Also updated the autoruler to take allies into account, making it
more useful for seeing real attack vectors and enabling you to
consider your allies as possible defense vectors.