Create a general filtering mechanism for all reports.
Change reports to be defined in terms of "stanzas" and "lines". A
report is now an array of either individual strings or subarrays
of strings called "Stanzas", and filtering works by always including
all the top level strings but including or excluding entire stanzas
based on whether or not any line in the stanza passes the applied
filter.

This means, for example,that you can have a top level structure to
define a table (for the accounting ledger for example) but the rows
of that table are isolated into stanzas so that they can be filtered.

In addition, you can have a lengthy multi-line result like in the
long fleet report, and filter out only those that relate to a star,
player, or specific string (like "Combat!") to then see the subset
of the report you're actually interested in.

Implement the filtered combat reports on top of this new structure
and pull the mechanism out into its own unit tested file for
documentation purposes. Next up, adding a UI to let the user custom
filter any report they are looking at.