Got distracted speeding up the ownership report.
The ownership report felt really laggy on the 64p game in combination
with filtering. Make the report itself require 50-100ms but ultimately
it is still laggy if you generate a giant report with thousands of
lines. It's probably good enough for now and long term fixes would
be to optimize Crus.format or to automatically truncate or filter
huge reports (e.g. by showing only the last 100 ticks).