Go back to slightly slower but more robust prior.
If there are scan logs with missing ticks, relying on the prior
coming from the patches would wind up observing the same events
multiple times, creating a worse ownership report. Go back to
keeping track internally at a slight cost so that the database
doesn't have to be perfect to give a nice looking report.