Move in memory append to after successful store.
One of the few ways I could see that it is possible to lose
a record in event storage is if the store request fails once
and then starts succeeding, causing the in memory store to
know about the unpersisted record and then skip over it forever
for future storage.

This change alters the order of operations so that if the persistent
storage fails, it won't be cached in RAM which will force it to
reattempt download and storage of the event.