Add data validation of event cache when the user views events.
Though hopefully the previous fix made it so that the event db
can't be corrupted, since we've now seen corruption in the wild
make it so that when the user views events there is a sanity
check on the database and a forced restore if a record is missing.

Some kind of server side logging will probably be needed for
αλφα Edition users so that occurrences of this problem can be
metered to prove that it is either fixed or that people are
hitting the restore case.