First version of menu access for NPA.
In an attempt to make the NPA functionality more discoverable and
the hotkeys more memorable, add a top level menu that lists all the
reports and their hotkeys. The top level menu itself can be invoked
with the 'm' hotkey, and if it is up, any menu item hotkey will
both execute and display the relevant report.

Thanks to Rob for proposing this FR.