Make time machine work when you've not shared your key.
A bug in prior versions relied on the idea that you would have
mailed your key to yourself for reference or to another player
to share, and if neither of those things had happened the time
machine would fail to pick up your scan data.

Change it so that any key you have merged is also in the list
of keys, and since generating a key merges it, the case where
you've never shared your key works.

This does still leave a bad edge case where you generate a key,
never share it, and then for some reason regenerate it. In that
case the older key data will no longer be found. I'll fix that
in a follow-up.