Automatically send a message with new API key.
Each time the user generates an API key, send yourself a message
with the API key so that it is findable by the code on other clients,
as well as known in case you regenerate it.

This fixes the edge case in the prior change by making it that
all keys are in your messages and repeatedly generating your key
won't cause the time machine to lose it.