Fix the map of undefined keys stacktrace in faReport.
Since I always seem to have some API keys I have seldom personally
seen one of the most commonly occurring stack traces in the wild: an
attempt to map over all API keys for the faReport.

Fix it so that the code doesn't execute if there are no API keys to
look at, as well as don't bother to execute it if FAs aren't even
enabled for the game.