Defend against a star not being a star.
The base game code has a bug when using the ruler that when the
user selects a fleet it sets universe.selectedStar to the fleet
instead of setting universe.selectedFleet, most likely because
using the ruler from fleets was a feature added in after the code
was initially written.

The autoruler assumed that the selectedStar was really a star and
was throwing (harmless) uncaught exceptions in this case. Fix it
to detect this case and avoid it.