Modify event overlap to account for ordering problem.
Frustratingly it turns out that if the timestamps on events are
identical, they aren't _always_ returned in the same order from
indexedDB and so they don't _always_ overlap in exactly the same
way.

Two possible fixes: either make the writes accept the idea that
duplicate records can happen (which would be harmless) or detect
out of order records with identical timestamps. For now I went with
the second as I like the tight constraint that exactly the correct
records are written to the db. This might be wrong, though, as we've
also seen cases of missing records rarely which could possibly be
caused by a similar ordering problem if the server isn't guaranteed
to send us all the events with the identical timestamp at the same
time. So in fact a fix that favours robustness might be better.
Something to worry about.