Partial fix for declaring war.
When you are one of the parties who was involved in the allied fleet
detection, you can use the warmap to be sure you're at war or not
without having to review the event log to know. Make the allied
fleet detection take this into account so that if you're at war,
that is correctly noted before the faReport data are used to decide
you're allied because you once shared a star.

Really the alliance detection also needs to watch the event log to
fix this properly.